Refactorings for constant folding
[ghstack-poisoned]
